Homology Functions

Regular Homology

ConleyDynamics.homology — Method

homology(lc::LefschetzComplex; [p::Int])

Compute the homology of a Lefschetz complex.

The homology is computed over the rationals (for p=0) or the finite field GF(p) (for prime p) and is returned as a vector betti of Betti numbers, where betti[k] is the Betti number in dimension k-1. If the Lefschetz complex boundary matrix already has been specialized to a field, the optional argument p can be omitted.

ConleyDynamics.relative_homology — Method

relative_homology(lc::LefschetzComplex,
                  subc::Union{Vector{Int},Vector{String}};
                  [p::Int])

Compute the relative homology of a Lefschetz complex with respect to a subcomplex.

The subcomplex is the closure of the cells in subc, which can be given either as indices or labels. The homology is computed over the rationals (for p=0) or the finite field GF(p) (for prime p) and is returned as a vector betti of Betti numbers, where betti[k] is the Betti number in dimension k-1. If the Lefschetz complex boundary matrix already has been specialized to a field, the optional argument p can be omitted.

Persistent Homology

ConleyDynamics.persistent_homology — Method

persistent_homology(lc::LefschetzComplex, filtration::Vector{Int};
                    [p::Int])

Complete the persistent homology of a Lefschetz complex filtration.

The function returns the two values

phsingles::Vector{Vector{Int}}

phpairs::Vector{Vector{Tuple{Int,Int}}}

It assumes that the order given by the filtration values is admissible, i.e., the permuted boundary matrix is strictly upper triangular. The persistence computation is performed over the finite field GF(p) (for prime p) or over the rationals (for p=0). The function returns the starting filtration values for infinite length persistence intervals in phsingles, and the birth- and death-filtration values for finite length persistence intervals in phpairs. If the Lefschetz complex boundary matrix already has been specialized to a field, then the optional argument p can be omitted.

Reduction Algorithm

ConleyDynamics.ph_reduce! — Method

ph_reduce!(matrix::SparseMatrix; [returnbasis=true])

Apply the persistence reduction algorithm to the matrix.

The function returns the values

phsingles::Vector{Vector{Int}}

phpairs::Vector{Vector{Tuple{Int,Int}}}

basis::SparseMatrix (if returnbasis=true)

It assumes that matrix is strictly upper triangular. The function returns the starting columns for infinite length persistence intervals in phsingles, and the birth- and death-columns for finite length persistence intervals in phpairs. If the optional argument returnbasis=true is given, then the function also returns the computed basis matrix B with reduced = matrix * B.