Sparse Matrix Functions

Access Functions

ConleyDynamics.sparse_get_entry — Method

value = sparse_get_entry(matrix::SparseMatrix, ri::Int, ci::Int)

Get the sparse matrix entry at location (ri,ci).

Base.getindex — Method

Base.getindex(matrix::SparseMatrix, ri::Int, ci::Int)

Get the sparse matrix entry at location (ri,ci).

ConleyDynamics.sparse_set_entry! — Method

sparse_set_entry!(matrix::SparseMatrix, ri::Int, ci::Int, val)

Set the sparse matrix entry at location (ri,ci) to val.

Base.setindex! — Method

Base.setindex!(matrix::SparseMatrix, val, ri::Int, ci::Int)

Set the sparse matrix entry at location (ri,ci) to val.

ConleyDynamics.sparse_get_column — Method

value = sparse_get_column(matrix::SparseMatrix, ci::Int)

Get the ci-th column of the sparse matrix.

ConleyDynamics.sparse_get_nz_column — Method

value = sparse_get_nz_column(matrix::SparseMatrix, ci::Int)

Get the row indizes for the nonzero entries in the ci-th column of the sparse matrix.

ConleyDynamics.sparse_minor — Method

smp = sparse_minor(sm::SparseMatrix, rvec::Vector{Int}, cvec::Vector{Int})

Create sparse submatrix by specifying the desired row and column indices.

Basic Functions

ConleyDynamics.sparse_size — Method

n = sparse_size(matrix::SparseMatrix, dim::Int)

Number of rows (dim=1) or columns (dim=2) of a sparse matrix.

ConleyDynamics.sparse_low — Method

n = sparse_low(matrix::SparseMatrix, col::Int)

Row index of the lowest nonzero matrix entry in column col.

ConleyDynamics.sparse_is_sut — Method

bool = sparse_is_sut(sm::SparseMatrix)

Check whether the sparse matrix is strictly upper triangular.

ConleyDynamics.sparse_identity — Method

sm = sparse_identity(n::Int, tone)

Create a sparse identity matrix with n rows and columns, and with diagonal entry tone.

ConleyDynamics.sparse_fullness — Method

sparse_fullness(sm::SparseMatrix)

Display the fullness of the sparse matrix sm.

ConleyDynamics.sparse_sparsity — Method

sparse_sparsity(sm::SparseMatrix)

Display the sparsity of the sparse matrix sm.

ConleyDynamics.sparse_show — Method

sparse_show(sm::SparseMatrix)

Display the sparse matrix sm.

ConleyDynamics.sparse_show — Method

sparse_show(sm::SparseMatrix{Int})

Display the sparse matrix sm.

Base.show — Method

Base.show(io::IO, sm::SparseMatrix)

Display the sparse matrix sm.

Base.show — Method

Base.show(io::IO, sm::SparseMatrix{Int})

Display the sparse matrix sm.

Elementary Matrix Operations

ConleyDynamics.sparse_add_column! — Method

sparse_add_column!(matrix::SparseMatrix, ci1::Int, ci2::Int, c)

Replace column[ci1] by column[ci1] + c * column[ci2].

ConleyDynamics.sparse_add_row! — Method

sparse_add_row!(matrix::SparseMatrix, ri1::Int, ri2::Int, c)

Replace row[ri1] by row[ri1] + c * row[ri2].

ConleyDynamics.sparse_permute — Method

smp = sparse_permute(sm::SparseMatrix, pr::Vector{Int}, pc::Vector{Int})

Create sparse matrix by permuting the row and column indices.

The vector pr describes the row permutation, and pc the column permutation.

ConleyDynamics.sparse_remove! — Method

sparse_remove!(matrix::SparseMatrix, ri::Int, ci::Int)

Remove the sparse matrix entry at location (ri,ci).

ConleyDynamics.sparse_multiply — Method

sm = sparse_multiply(A::SparseMatrix, B::SparseMatrix)

Multiply two sparse matrices.

Exceptions are raised if the matrix product is not defined or the entry types do not match.

Base.:* — Method

sm = Base.:*(A::SparseMatrix, B::SparseMatrix)

Multiply two sparse matrices.

Exceptions are raised if the matrix product is not defined or the entry types do not match.

Conversion Functions

ConleyDynamics.sparse_from_full — Method

sm = sparse_from_full(matrix::Matrix{Int})

Create sparse matrix from full matrix.

ConleyDynamics.full_from_sparse — Method

fm = full_from_sparse(sm::SparseMatrix)

Create full matrix from sparse matrix.

ConleyDynamics.sparse_from_lists — Method

sm = sparse_from_lists(nr, nc, tzero, tone, r, c, vals)

Create sparse matrix from lists describing the entries.

The vectors r, c, and vals have to have the same length and the matric has entry vals[k] at (r[k],c[k]). Zero entries will be ignored, and multiple entries for the same matrix position raise an error.

The input arguments have the following meaning:

nr::Int: Number of rows

nc::Int: Number of columns

tzero::T: Number 0 of type T

tone::T: Number 1 of type T

r::Vector{Int}: Vector of row indices

c::Vector{Int}: Vector of column indices

vals::Vector{T}: Vector of matrix entries

ConleyDynamics.lists_from_sparse — Method

nr, nc, tzero, tone, r, c, vals = lists_from_sparse(sm::SparseMatrix)

Create list representation from sparse matrix.

The output variables are exactly what is needed to create a sparse matrix object using sparse_from_lists.