Sparse Matrices

While Julia provides a data structure for sparse matrix computations, the employed design decisions make it difficult to use this implementation for computations over finite fields. This is mainly due to the fact that in the Julia implementation, it is assumed that one can determine the zero and one elements from the data type alone. However, a finite field data type generally also depends on additional parameters, such as the characteristic of the field.

Since the algorithms underlying ConleyDynamics.jl only require basic row and column operations, a specialized sparse matrix implementation is provided in the package. It is briefly described in the following.

Sparse Matrix Format

ConleyDynamics.SparseMatrix — Type

SparseMatrix{T}

Composite data type for a sparse matrix with entries of type T.

The struct has the following fields:

const nrow::Int: Number of rows

const ncol::Int: Number of columns

const char::Int: Characteristic of type T

const zero::T: Number 0 of type T

const one::T: Number 1 of type T

entries::Vector{Vector{T}}: Matrix entries corresponding to columns

columns::Vector{Vector{Int}}: column[k] points to nonzero entries in column k

rows::Vector{Vector{Int}}: rows[k] points to nonzero entries in the k-th row

Creating Sparse Matrices

sparse_from_full

full_from_sparse

sparse_from_lists

lists_from_sparse

sparse_identity

Sparse Matrix Access

sparse_get_entry

sparse_set_entry!

sparse_get_column

sparse_get_nz_column

sparse_minor

You can also use y = A[i,j] and A[i,j] = val.

Sparse Matrix Information

sparse_size

sparse_low

sparse_is_sut

sparse_fullness

sparse_sparsity

sparse_show

Elementary Matrix Operations

sparse_add_column!

sparse_add_row!

sparse_permute

sparse_remove!

sparse_multiply

You can also use A*A to compute the product of sparse matrices.