Conley Theory Functions

Conley Index Computations

ConleyDynamics.conley_index — Function

conley_index(lc::LefschetzComplex, subcomp::Vector{String})

Determine the Conley index of a Lefschetz complex subset.

The function raises an error if the subset subcomp is not locally closed. The computations are performed over the field associated with the Lefschetz complex boundary matrix.

conley_index(lc::LefschetzComplex, subcomp::Vector{Int})

Determine the Conley index of a Lefschetz complex subset.

The function raises an error if the subset subcomp is not locally closed. The computations are performed over the field associated with the Lefschetz complex boundary matrix.

ConleyDynamics.morse_sets — Function

morse_sets(lc::LefschetzComplex, mvf::CellSubsets; poset::Bool=false)

Find the nontrivial Morse sets of a multivector field on a Lefschetz complex.

The input argument lc contains the Lefschetz complex, and mvf describes the multivector field. The function returns the nontrivial Morse sets as a Vector{Vector{Int}}. If the optional argument poset=true1 is added, then the function returns both the Morse sets and the adjacency matrix of the Hasse diagram of the underlying poset.

ConleyDynamics.invariant_hull — Function

invariant_hull(lc::LefschetzComplex, mvf::CellSubsets,
               ms::CellSubsets)

Find the isolated invariant hull of a collection of Morse sets.

The input argument lc contains the Lefschetz complex, and mvf describes the multivector field. The collection of Morse sets are contained inms. All of these sets have to be Morse sets in the sense of being strongly connected components of the flow graph. (For runtime issues, this will not be checked!) In other words, the sets in ms have to be determined using the function morse_sets!

The function returns the smallest isolated invariant set which contains the Morse sets and their connections as a Vector{Int}.

ConleyDynamics.restrict_dynamics — Function

restrict_dynamics(lc::LefschetzComplex, mvf::CellSubsets, lcsub::Cells)

Restrict a multivector field to a Lefschetz subcomplex.

For a given multivector field mvf on a Lefschetz complex lc, and a subcomplex which is given by the locally closed set represented by lcsub, create the associated Lefschetz subcomplex lcreduced and the induced multivector field mvfreduced on the subcomplex. The multivectors of the new multivector field are the intersections of the original multivectors and the subcomplex.

Connection Matrix Computation

ConleyDynamics.connection_matrix — Function

connection_matrix(lc::LefschetzComplex, mvf::CellSubsets;
                  [returnbasis::Bool])

Compute a connection matrix for the multivector field mvf on the Lefschetz complex lc over the field associated with the Lefschetz complex boundary matrix.

The function returns an object of type ConleyMorseCM. If the optional argument returnbasis::Bool=true is given, then the function also returns a dictionary which gives the basis for the connection matrix columns in terms of the original labels.

ConleyDynamics.cm_reduce! — Function

cm_reduce!(matrix::SparseMatrix, psetvec::Vector{Int};
           [returnbasis::Bool],[returntm::Bool])

Compute the connection matrix.

Assumes that matrix is upper triangular and filtered according to psetvec. Modifies the argument matrix.

Return values:

cmatrix: Connection matrix

cmatrix_cols: Columns of the connection matrix in the boundary

basisvecs (optional): If the argument returnbasis=true is given, this returns information about the computed basis. The k-th entry of basisvecs is a vector containing the columns making up the k-th basis vector, which corresponds to column cmatrix_cols[k].

tmatrix (optional): If the argument returntm=true is given in addition to returnbasis=true, then instead of basisvecs the function returns the complete transformation matrix. In this case, basicvecs is not returned.

Poset Order Functions

ConleyDynamics.admissible_order — Function

admissible_order(bndmatrix::SparseMatrix, mvf::Vector{Vector{Int}})

Find an admissible order based on the boundary matrix and the multivector field.

The vector sccnumber contains the strongly connected component for each column, and the matrix admibnd is the reordered boundary matrix corresponding to the order given by admiorder.

Return values:

admiorder: Admissible order

admibnd: Reordered boundary matrix

sccnumber: Strongly connected component index of column

scc: Strongly connected components

ConleyDynamics.renumber_poset! — Function

renumber_poset!(poset::Vector{Int})

Renumber the poset given by the increasing integer vector poset.

Matrix Helper Functions

ConleyDynamics.cm_columns — Function

cm_columns(matrix::SparseMatrix, psetvec::Vector{Int})

Create a vector of column indices for the connection matrix.

ConleyDynamics.homogeneous_columns — Function

homogeneous_columns(matrix::SparseMatrix, psetvec::Vector{Int})

Determine which columns of matrix are homogenenous columns.

ConleyDynamics.is_homogeneous — Function

is_homogeneous(matrix::SparseMatrix, psetvec::Vector{Int}, cindex::Int)

Decide whether a column of matrix is homogeneous.

ConleyDynamics.target_columns — Function

target_columns(matrix::SparseMatrix, psetvec::Vector{Int})

Determine which columns of matrix are target columns.