Connection Matrix Functions

Connection Matrix Computation

ConleyDynamics.connection_matrix — Method

cm = connection_matrix(lc, mvf; p=2)
cm, cmbasis = connection_matrix(lc, mvf; p=2, returnbasis=true)

Compute a connection matrix for the multivector field mvf on the Lefschetz complex lc over a finite field with p elements.

The arguments are typed as lc::LefschetzComplex and mvf::Vector{Vector{Int}}, and the return object is of type ConleyMorseCM. If the optional argument returnbasis::Bool=true is given, then the function returns a dictionary which gives the basis for the connection matrix columns in terms of the original labels. If p is omitted, then p=2 is used.

ConleyDynamics.connection_matrix — Method

cm = connection_matrix(lc, mvf; p=2)
cm, cmbasis = connection_matrix(lc, mvf; p=2, returnbasis=true)

Compute a connection matrix for the multivector field mvf on the Lefschetz complex lc over a finite field with p elements.

The arguments are typed as lc::LefschetzComplex and mvf::Vector{Vector{String}}, and the return object is of type ConleyMorseCM. If the optional argument returnbasis::Bool=true is given, then the function returns a dictionary which gives the basis for the connection matrix columns in terms of the original labels. If p is omitted, then p=2 is used.

ConleyDynamics.cm_create! — Method

cmatrix, cmatrix_cols = cm_create!(matrix, psetvec)
cmatrix, cmatrix_cols, basisvecs = cm_create!(matrix, psetvec;
                                              returnbasis=true)

Compute the connection matrix.

Assumes that matrix is upper triangular and filtered according to psetvec. Modifies the argument matrix. If the optional argument returnbasis=true is given, then the function also returns information about the computed basis. The k-th entry of basisvecs is a vector containing the columns making up the k-th basis vector, which corresponds to column cmatrix_cols[k].

ConleyDynamics.cm_create! — Method

cmatrix, cmatrix_cols = cm_create!(matrix, psetvec)
cmatrix, cmatrix_cols, basisvecs = cm_create!(matrix, psetvec;
                                              returnbasis=true)

Compute the connection matrix.

Assumes that matrix is upper triangular and filtered according to psetvec. Modifies the argument matrix. If the optional argument returnbasis=true is given, then the function also returns information about the computed basis. The k-th entry of basisvecs is a vector containing the columns making up the k-th basis vector, which corresponds to column cmatrix_cols[k].

Poset Order Functions

ConleyDynamics.admissible_order — Method

admiorder, admibnd, sccnumber, scc = admissible_order(bndmatrix, mvf)

Find an admissible order based on the boundary matrix and the multivector field.

The vector sccnumber contains the strongly connected component for each column, and the matrix admibnd is the reordered boundary matrix corresponding to the order given by admiorder.

ConleyDynamics.admissible_order — Method

admiorder, admibnd, sccnumber, scc = admissible_order(bndmatrix, mvf)

Find an admissible order based on the boundary matrix and the multivector field.

The vector sccnumber contains the strongly connected component for each column, and the matrix admibnd is the reordered boundary matrix corresponding to the order given by admiorder.

ConleyDynamics.renumber_poset! — Method

renumber_poset!(poset)

Renumber the poset given by the increasing integer vector poset.

Sparse Matrix Helper Functions

ConleyDynamics.cm_columns — Method

ccols = cm_columns(matrix::SparseMatrix, psetvec)

Create a vector of column indices for the connection matrix.

ConleyDynamics.homogeneous_columns — Method

hcols = homogeneous_columns(matrix::SparseMatrix, psetvec)

Determine which columns of matrix are homogenenous columns.

ConleyDynamics.is_homogeneous — Method

bool = is_homogeneous(matrix::SparseMatrix, psetvec, cindex)

Decide whether a column of matrix is homogeneous.

ConleyDynamics.target_columns — Method

tcols = target_columns(matrix::SparseMatrix, psetvec)

Determine which columns of matrix are target columns.

Full Matrix Helper Functions

ConleyDynamics.cm_columns — Method

ccols = cm_columns(matrix, lowvec, psetvec)

Create a vector of column indices for the connection matrix.

ConleyDynamics.homogeneous_columns — Method

hcols = homogeneous_columns(matrix, lowvec, psetvec)

Determine which columns of matrix are homogenenous columns.

ConleyDynamics.is_homogeneous — Method

bool = is_homogeneous(matrix, lowvec, psetvec, cindex)

Decide whether a column of matrix is homogeneous.

ConleyDynamics.target_columns — Method

tcols = target_columns(matrix, lowvec, psetvec)

Determine which columns of matrix are target columns.

ConleyDynamics.update_low! — Method

update_low!(matrix, lowvec; startindex=1)

Compute the indices of the lowest nonzero entries in each column of matrix.

The indices are stored in lowvec. If startindex > 1, then only the values starting with column startindex are updated.