Conley Theory Functions

Conley Index Computations

ConleyDynamics.conley_index — Method

conley_index(lc::LefschetzComplex, subcomp::Vector{Int};
             p::Int)

Determine the Conley index of a Lefschetz complex subset.

The function raises an error if the subset subcomp is not locally closed. The optional parameter p specifies the field characteristic for the homology computation.

ConleyDynamics.conley_index — Method

conley_index(lc::LefschetzComplex, subcomp::Vector{String};
             p::Int)

Determine the Conley index of a Lefschetz complex subset.

The function raises an error if the subset subcomp is not locally closed. The optional parameter p specifies the field characteristic for the homology computation.

Connection Matrix Computation

ConleyDynamics.connection_matrix — Method

connection_matrix(lc::LefschetzComplex, mvf::MultiVectorField;
                  [p::Int,] [returnbasis::Bool])

Compute a connection matrix for the multivector field mvf on the Lefschetz complex lc over a finite field with p elements, or over the rationals.

The function returns an object of type ConleyMorseCM. If the optional argument returnbasis::Bool=true is given, then the function also returns a dictionary which gives the basis for the connection matrix columns in terms of the original labels. If p is omitted, then the Lefschetz complex boundary has to has been specified over a field. If the boundary matrix is an integer matrix, p has to be chosen.

ConleyDynamics.cm_reduce! — Method

cm_reduce!(matrix::SparseMatrix, psetvec::Vector{Int};
           [returnbasis::Bool])

Compute the connection matrix.

Assumes that matrix is upper triangular and filtered according to psetvec. Modifies the argument matrix.

Return values:

cmatrix: Connection matrix

cmatrix_cols: Columns of the connection matrix in the boundary

basisvecs (optional): If the argument returnbasis=true is given, this returns information about the computed basis. The k-th entry of basisvecs is a vector containing the columns making up the k-th basis vector, which corresponds to column cmatrix_cols[k].

Poset Order Functions

ConleyDynamics.admissible_order — Method

admissible_order(bndmatrix::SparseMatrix, mvf::Vector{Vector{Int}})

Find an admissible order based on the boundary matrix and the multivector field.

The vector sccnumber contains the strongly connected component for each column, and the matrix admibnd is the reordered boundary matrix corresponding to the order given by admiorder.

Return values:

admiorder: Admissible order

admibnd: Reordered boundary matrix

sccnumber: Strongly connected component index of column

scc: Strongly connected components

ConleyDynamics.renumber_poset! — Method

renumber_poset!(poset::Vector{Int})

Renumber the poset given by the increasing integer vector poset.

Matrix Helper Functions

ConleyDynamics.cm_columns — Method

cm_columns(matrix::SparseMatrix, psetvec::Vector{Int})

Create a vector of column indices for the connection matrix.

ConleyDynamics.homogeneous_columns — Method

homogeneous_columns(matrix::SparseMatrix, psetvec::Vector{Int})

Determine which columns of matrix are homogenenous columns.

ConleyDynamics.is_homogeneous — Method

is_homogeneous(matrix::SparseMatrix, psetvec::Vector{Int}, cindex::Int)

Decide whether a column of matrix is homogeneous.

ConleyDynamics.target_columns — Method

target_columns(matrix::SparseMatrix, psetvec::Vector{Int})

Determine which columns of matrix are target columns.