Plotting Functions

Visualizing Simplicial Complexes and Forman Vector Fields

ConleyDynamics.plot_planar_simplicial — Method

plot_planar_simplicial(sc::LefschetzComplex,
                       coords::Vector{<:Vector{<:Real}},
                       fname::String;
                       mvf::MultiVectorField,
                       labeldir::Vector{<:Real},
                       hfac::Real,
                       vfac::Real)

Create an svg image of a planar simplicial complex, and if specified, a Forman vector field on it.

The vector coords contains coordinates for every one of the vertices of the simplicial complex sc. The image will be saved in the file with name fname. The optional vector labeldir contains directions for the vertex labels. These have to be reals between 0 and 4, with 0,1,2,3 corresponding to E,N,W,S. If mvf is specified and is a Forman vector field, then this Forman vector field is drawn as well. The optional constants hfac and vfac contain the horizontal and vertical scale vectors.