Utility Functions

Simplicial Complexes

ConleyDynamics.create_simplicial_complex — Method

lc = create_simplicial_complex(labels::Vector{String},
                               simplices::Vector{Vector{Int}})

Initialize a Lefschetz complex from a simplicial complex.

The vector labels contains a label for every vertex, while simplices contains all the highest-dimensional simplices necessary to define the simplicial complex. Every simplex is represented as a vector of Int, with entries corresponding to the vertex indices.

Note that the labels all have to have the same character length!

ConleyDynamics.create_simplicial_complex — Method

lc = create_simplicial_complex(labels::Vector{String},
                               simplices::Vector{Vector{String}})

Initialize a Lefschetz complex from a simplicial complex.

The vector labels contains a label for every vertex, while simplices contains all the highest-dimensional simplices necessary to define the simplicial complex.

ConleyDynamics.create_simplicial_rectangle — Method

create_simplicial_rectangle(nx::Int, ny::Int)

Create a simplicial complex covering a rectangle in the plane.

The rectangle is given by the subset [0,nx] x [0,ny] of the plane. Each unit square is represented by four triangles, which meet in the center point of the square. Labels have the following meaning:

The label XXXYYYb corresponds to the point (XXX, YYY).

The label XXXYYYc corresponds to (XXX + 1/2, YYY + 1/2).

The number of characters in XXX and YYY matches the number of digits of the larger number of nx and ny.

ConleyDynamics.convert_simplices — Method

newsimplicies = convert_simplices(simplices::Vector{Vector{Int}},
                                  labels::Vector{String})

Convert list of simplices from index form to label form.

ConleyDynamics.convert_simplices — Method

newsimplicies = convert_simplices(simplices::Vector{Vector{String}},
                                  labels::Vector{String})

Convert list of simplices from label form to index form.

Lefschetz complexes

ConleyDynamics.lefschetz_closure — Method

c = lefschetz_closure(lc::LefschetzComplex, subcomp::Vector{Int})

Compute the closure of a Lefschetz complex subset.

ConleyDynamics.lefschetz_closure — Method

c = lefschetz_closure(lc::LefschetzComplex, subcomp::Vector{String})

Compute the closure of a Lefschetz complex subset.

ConleyDynamics.lefschetz_subcomplex — Method

lefschetz_subcomplex(lc::LefschetzComplex, subcomp::Vector{Int})

Extract a closed subcomplex from a Lefschetz complex. The subcomplex is the closure of the collection of cells given in subcomp.

ConleyDynamics.lefschetz_subcomplex — Method

lefschetz_subcomplex(lc::LefschetzComplex, subcomp::Vector{String})

Extract a closed subcomplex from a Lefschetz complex. The subcomplex is the closure of the collection of cells given in subcomp.

ConleyDynamics.permute_lefschetz_complex — Method

lc2 = permute_lefschetz_complex(lc::LefschetzComplex,
                                permutation::Vector{Int})

Permute the indices of a Lefschetz complex.

The vector permutation contains a permutation of the indices for the given Lefschetz complex lc. If no permutation is specified, or if the length of the vector is not correct, then a randomly generated one will be used.

ConleyDynamics.convert_lefschetz_sparse — Method

lc2 = convert_lefschetz_sparse(lc::LefschetzComplex)

Convert the boundary matrix of a Lefschetz complex to sparse.

Multivector Fields

ConleyDynamics.convert_mvf — Method

newmvf = convert_mvf(mvf::Vector{Vector{Int}}, lc::LefschetzComplex)

Convert multivector field mvf on the Lefschetz complex lc from index form to label form.

ConleyDynamics.convert_mvf — Method

newmvf = convert_mvf(mvf::Vector{Vector{String}}, lc::LefschetzComplex)

Convert multivector field mvf on the Lefschetz complex lc from label form to index form.

Matrix Conversions

ConleyDynamics.convert_matrix_gfp — Method

gfpmatrix = convert_matrix_gfp(matrix::Matrix{Int}, p::Int)

Convert an integer matrix to a finite field matrix over GF(p). For the choice p=0 the rationals are used.

ConleyDynamics.convert_matrix_gfp — Method

gfpmatrix = convert_matrix_gfp(matrix::SparseMatrix{Int}, p::Int)

Convert a sparse integer matrix to a finite field sparse matrix over GF(p). For the choice p=0 the rationals are used.

ConleyDynamics.convert_matrix_int — Method

intmatrix = convert_matrix_int(matrix::Matrix)

Convert a matrix over the finite field GF(p) to an integer matrix.